Trait diem_metrics_core::Encoder

pub trait Encoder {
    pub fn encode<W>(&self, &[MetricFamily], &mut W) -> Result<(), Error>
    where
        W: Write;

    pub fn format_type(&self) -> &str;
}

An interface for encoding metric families into an underlying wire protocol.

Required methods

pub fn encode<W>(&self, &[MetricFamily], &mut W) -> Result<(), Error> where
W: Write,

encode converts a slice of MetricFamily proto messages into target format and writes the resulting lines to writer. It returns the number of bytes written and any error encountered. This function does not perform checks on the content of the metric and label names, i.e. invalid metric or label names will result in invalid text format output.

pub fn format_type(&self) -> &str

format_type returns target format.

Loading content...